requirements.txt
gradio==4.26.0
moviepy==2.0.0.dev2
pyroomacoustics
noisereduce
opencv-python
#diffusers
#edge_tts==6.1.8
edge_tts
transformers
sentencepiece
accelerate
safetensors
lac
paddlepaddle==2.5.2

pyinstaller
sacremoses
#加载3D
bpy

#cn2an
numpy
scipy
tensorboard
librosa==0.9.2
numba==0.56.4
pytorch-lightning
ffmpeg-python
onnxruntime
tqdm
funasr==1.0.0
pypinyin
pyopenjtalk
g2p_en
modelscope==1.10.0
sentencepiece
transformers
chardet
PyYAML
psutil
jieba_fast
LangSegment>=0.2.0
Faster_Whisper
wordsegment


#openai>=1.0

#https://llama.family/docs/api


#pip install resemble-enhance --upgrade --pre


#4.25.3
protobuf==3.20.1


celluloid==0.2.0
#deepspeed
matplotlib==3.8.1
omegaconf==2.3.0
pandas==2.1.3
ptflops==0.7.1.2
rich==13.7.0
soundfile
resampy==0.4.2
tabulate==0.8.10
#async_io==0.3.16
#deepspeed

celluloid==0.2.0
matplotlib==3.8.1
omegaconf==2.3.0
pandas==2.1.3
ptflops==0.7.1.2
rich==13.7.0
soundfile==0.12.1
resampy==0.4.2
tabulate==0.8.10
loguru


# pip install torch==2.1.0 torchvision==0.16.0 torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu121
#@echo Instaling deepspeed for python 3.10.x or 3.11 and CUDA 11.8
# python deepspeed_installer.py

#spaces


# pip install -q -U google-generativeai
# pip install httpx[socks]

#pip3 install -U xformers --index-url https://download.pytorch.org/whl/cu118

#llama3

bitsandbytes
flash_attn
#pip install flash-attn --no-build-isolation


diffusers==0.28.0

#diffusers==0.27.2